del 3

+++ button_actions.py
@@ -1,6 +1,7 @@
+
 from pony.orm import *
 
-import Core.database as database
+import database as database
 import datetime
 
 db = Database()
@@ -78,6 +79,8 @@ def book_media(bot, update):
         session = database.RegistrySession[telegram_id]
         media = list(database.Media.select())[session.media_c]
 
+
+
         # If media can't be booked, then reject booking
         if media.availability == 0:
             bot.edit_message_text(text="🤦🏻‍♂️ Media can't be booked. This item is not available now",
@@ -126,11 +129,30 @@ def book_media(bot, update):
                               message_id=query.message.message_id)
 
 
-"""
-def make_return_request(media, patron, issue_date)
 
-This function generates expiry date based on type of media and user.
-"""
+def add_in_line(bot, update):
+    query = update.callback_query
+    telegram_id = query.message.chat_id
+    session = database.RegistrySession[telegram_id]
+    media = list(database.Media.select())[session.media_c]
+    user = database.User[telegram_id]
+    a = list(database.Log.select(
+            lambda c: c.libID == telegram_id and c.mediaID.startswith(str(media.mediaID))))
+    if not user.is_in_line(media) and len(list(database.Log.select(
+            lambda c: c.libID == telegram_id and c.mediaID.startswith(str(media.mediaID))))) == 0:
+        user.add_to_queue(media)
+        bot.edit_message_text(text="You have been successfully added to the line!",
+                              chat_id=query.message.chat_id,
+                              message_id=query.message.message_id)
+    else:
+        bot.edit_message_text(text="Sorry, mate, but you are already in line!",
+                              chat_id=query.message.chat_id,
+                              message_id=query.message.message_id)
+    """
+    def make_return_request(media, patron, issue_date)
+    
+    This function generates expiry date based on type of media and user.
+    """
 
 
 def make_return_request(bot, update, copy_id):
@@ -150,22 +172,41 @@ def accept_return(bot, update, request_id):
     request = database.ReturnRequest[request_id]
     copy_id = request.copyID
     user_id = request.telegramID
+    media = database.MediaCopies.get(copyID=copy_id).mediaID
 
     #   Setting log.return to 1
     log_record = database.Log.select(lambda c: c.mediaID == request.copyID and not c.returned)
     log_record.returned = True
 
-    #   Correcting user medias
-    media_copy = database.MediaCopies.get(copyID=copy_id)
-    media_copy.available = True
 
-    request.delete()
+    if media.queue.is_empty():
+        #   Setting log.return to 1
+        log_record = database.Log.select(lambda c: c.mediaID == request.copyID and not c.returned)
+        log_record.returned = True
 
-    commit()
-    bot.send_message(text="Media %s has been successfully returned" % copy_id, chat_id=user_id)
-    bot.edit_message_text(text="Media %s has been successfully returned" % copy_id,
-                          message_id=update.callback_query.message.message_id,
-                          chat_id=update.callback_query.message.chat_id)
+        #   Correcting user medias
+        media_copy = database.MediaCopies.get(copyID=copy_id)
+        media_copy.available = True
+
+        request.delete()
+
+        commit()
+        bot.send_message(text="Media %s has been successfully returned" % copy_id, chat_id=user_id)
+        bot.edit_message_text(text="Media %s has been successfully returned" % copy_id,
+                              message_id=update.callback_query.message.message_id,
+                              chat_id=update.callback_query.message.chat_id)
+    else:
+        user = media.pop().user
+        # Making a record in log
+        database.Log(libID=user.telegramID, mediaID=copy_id,
+                     expiry_date=generate_expiry_date(media, user, datetime.datetime.now()))
+        bot.send_message(text="Media %s has been successfully returned" % copy_id, chat_id=user_id)
+        bot.edit_message_text(text="Media %s has been successfully returned" % copy_id,
+                              message_id=update.callback_query.message.message_id,
+                              chat_id=update.callback_query.message.chat_id)
+        bot.send_message(text="Dear %s, media #%s is available now! You can take it from library, time starts now!" % (user.name, copy_id),
+                         chat_id=user.telegramID)
+
 
 
 def reject_return(bot, update, request_id):
@@ -192,15 +233,18 @@ def ask_for_return(bot, update, copy_id, user_id):
     bot.send_message(
         text="""Hello, %s! 
 You recently took a book %s by %s (%s). Library and librarians need it ASAP. Could you please bring it back? Thank you!""" % (
-        user.name,
-        abstract_media.name,
-        abstract_media.authors,
-        copy_id),
+
+            user.name,
+            abstract_media.name,
+            abstract_media.authors,
+            copy_id),
+
 
         chat_id=user_id)
     bot.edit_message_text(text="Message to %s(@%s) has been sent" % (user.name, user.alias),
                           message_id=update.callback_query.message.message_id,
                           chat_id=update.callback_query.message.chat_id)
+
 
 """
 def generate_expiry_date(self, media, patron, issue_date)
